+++ R/get_predicted.R
@@ -1,23 +1,12 @@
 #' Model Predictions (robust)
 #'
-#' The \code{get_predicted()} function is a robust, flexible and user-friendly alternative to base R \code{\link{predict}} function. Additional features and advantages include availability of uncertainty intervals (CI), bootstrapping, a more intuitive API and the support of more models than base R's \code{predict}. However, although the interface are simplified, it is still very important to read the documentation of the arguments. This is because making "predictions" (a lose term for a variety of things) is a non-trivial process, with lots of caveats and complications. Read the \code{Details} section for more information.
+#' The \code{get_predicted()} function is a robust, flexible and user-friendly alternative to base R \code{\link{predict}} function. Additional features and advantages include availability of uncertainty intervals (CI), bootstrapping, a more intuitive API and the support of more models than base R's \code{predict} function. However, although the interface are simplified, it is still very important to read the documentation of the arguments. This is because making "predictions" (a lose term for a variety of things) is a non-trivial process, with lots of caveats and complications. Read the \code{Details} section for more information.
 #'
 #' @param x A statistical model (can also be a data.frame, in which case the
 #'   second argument has to be a model).
 #' @param data An optional data frame in which to look for variables with which
 #'   to predict. If omitted, the data used to fit the model is used.
-#' @param predict Can be \code{"link"}, \code{"expectation"} (default), or
-#'   \code{"prediction"}. This modulates the scale of the output as well as the
-#'   type of certainty interval. More specifically, \code{"link"} gives an
-#'   output on the link-scale (for logistic models, that means the log-odds
-#'   scale) with a confidence interval (CI). \code{"expectation"} (default) also
-#'   returns confidence intervals, but this time the output is on the response
-#'   scale (for logistic models, that means probabilities). Finally,
-#'   \code{"predict"} also gives an output on the response scale, but this time
-#'   associated with a prediction interval (PI), which is larger than a
-#'   confidence interval (though it mostly make sense for linear models). Read
-#'   more about in the \strong{Details} section below. \code{"relation"} is
-#'   also accepted as a (deprecated) alias for \code{"expectation"}.
+#' @param predict Can be \code{"link"}, \code{"expectation"} (default), \code{"prediction"}, or \code{"response"}. You can see these 4 options for predictions as on a gradient from "close to the model" to "close to the response data". More specifically, the \code{predict} argument modulates two things; the scale of the output as well as the type of certainty interval (see the details and examples). More specifically, \code{"link"} returns predictions on the model's link-scale (for logistic models, that means the log-odds scale) with a confidence interval (CI). \code{"expectation"} (default) also returns confidence intervals, but this time the output is on the response scale (for logistic models, that means probabilities). \code{"predict"} also gives an output on the response scale, but this time associated with a prediction interval (PI), which is larger than a confidence interval (though it mostly make sense for linear models). Finally, \code{"response"} only differs from the previous option for binomial models where it additionally transforms the predictions into the original response's type (for instance, to a factor). Read more about in the \strong{Details} section below.
 #' @param iterations For Bayesian models, this corresponds to the number of
 #'   posterior draws. If \code{NULL}, will return all the draws (one for each
 #'   iteration of the model). For frequentist models, if not \code{NULL}, will
@@ -44,10 +33,10 @@
 #'   columns and observations are rows) can be accessed via \code{as.data.frame}.
 #'
 #' @details
-#' The \code{predict} argument jointly modulates two separate concepts, the
-#' \strong{scale} and the \strong{uncertainty interval}.
+#' In \code{insight::get_predicted()}, the \code{predict} argument jointly
+#' modulates two separate concepts, the \strong{scale} and the \strong{uncertainty interval}.
 #'
-#' \subsection{Confidence Interval vs. Prediction Interval)}{
+#' \subsection{Confidence Interval (CI) vs. Prediction Interval (PI))}{
 #' \itemize{
 #'   \item \strong{Linear models} - \code{lm()}: For linear models, Prediction
 #'   intervals (\code{predict = "prediction"}) show the range that likely
@@ -72,17 +61,21 @@
 #' Having the output is on the scale of the response variable is arguably the
 #' most convenient to understand and visualize the relationships. If on the
 #' link-scale, no transformation is applied and the values are on the scale of
-#' the model's predictors. For instance, for a logistic model, the response
+#' the model. For instance, for a logistic model, the response
 #' scale corresponds to the predicted probabilities, whereas the link-scale
-#' makes predictions of log-odds (probabilities on the logit scale).
+#' makes predictions of log-odds (probabilities on the logit scale). Note that,
+#' when \code{predict = "response"}, the probabilities are rounded (so that the
+#' prediction corresponds to the most likely outcome).
 #' }
 #'
 #' @examples
 #' data(mtcars)
 #' x <- lm(mpg ~ cyl + hp, data = mtcars)
+#'
 #' predictions <- get_predicted(x)
 #' predictions
 #'
+#' # Options and methods ---------------------
 #' get_predicted(x, predict = "prediction")
 #'
 #' # Get CI
@@ -91,6 +84,29 @@
 #' # Bootstrapped
 #' as.data.frame(get_predicted(x, iterations = 4))
 #' summary(get_predicted(x, iterations = 4)) # Same as as.data.frame(..., keep_iterations = F)
+#'
+#' # Different predicttion types ------------------------
+#' data(iris)
+#' data <- droplevels(iris[1:100, ])
+#'
+#' # Fit a logistic model
+#' x <- glm(Species ~ Sepal.Length, data = data, family = "binomial")
+#'
+#' # Expectation (default): response scale + CI
+#' pred <- get_predicted(x, predict = "expectation")
+#' head(as.data.frame(pred))
+#'
+#' # Prediction: response scale + PI
+#' pred <- get_predicted(x, predict = "prediction")
+#' head(as.data.frame(pred))
+#'
+#' # Link: link scale + CI
+#' pred <- get_predicted(x, predict = "link")
+#' head(as.data.frame(pred))
+#'
+#' # Response: response "type" + PI
+#' pred <- get_predicted(x, predict = "response")
+#' head(as.data.frame(pred))
 #' @export
 get_predicted <- function(x, ...) {
   UseMethod("get_predicted")
@@ -147,7 +163,7 @@ get_predicted.data.frame <- function(x, data = NULL, verbose = TRUE, ...) {
 #' @export
 get_predicted.lm <- function(x,
                              data = NULL,
-                             predict = c("expectation", "link", "prediction", "relation"),
+                             predict = c("expectation", "link", "prediction", "response", "relation"),
                              iterations = NULL,
                              verbose = TRUE,
                              ...) {
@@ -256,7 +272,7 @@ get_predicted.merMod <- get_predicted.lmerMod
 #' @export
 get_predicted.glmmTMB <- function(x,
                                   data = NULL,
-                                  predict = c("expectation", "link", "prediction", "relation"),
+                                  predict = c("expectation", "link", "prediction", "response", "relation"),
                                   ci = 0.95,
                                   include_random = TRUE,
                                   iterations = NULL,
@@ -332,7 +348,7 @@ get_predicted.glmmTMB <- function(x,
 #' @export
 get_predicted.gam <- function(x,
                               data = NULL,
-                              predict = c("expectation", "link", "prediction", "relation"),
+                              predict = c("expectation", "link", "prediction", "response", "relation"),
                               ci = 0.95,
                               include_random = TRUE,
                               include_smooth = TRUE,
@@ -422,7 +438,7 @@ get_predicted.list <- get_predicted.gam # gamm4
 #' @export
 get_predicted.stanreg <- function(x,
                                   data = NULL,
-                                  predict = c("expectation", "link", "prediction", "relation"),
+                                  predict = c("expectation", "link", "prediction", "response", "relation"),
                                   iterations = NULL,
                                   include_random = TRUE,
                                   include_smooth = TRUE,
@@ -579,10 +595,12 @@ get_predicted.faMain <- function(x, data = NULL, ...) {
 }
 
 
+# -------------------------------------------------------------------------
+
 
 .get_predicted_args <- function(x,
                                 data = NULL,
-                                predict = c("expectation", "link", "prediction", "relation"),
+                                predict = c("expectation", "link", "prediction", "response", "response", "relation"),
                                 include_random = TRUE,
                                 include_smooth = TRUE,
                                 ci = 0.95,
@@ -591,7 +609,7 @@ get_predicted.faMain <- function(x, data = NULL, ...) {
                                 ...) {
 
   # Sanitize input
-  predict <- match.arg(predict, choices = c("expectation", "link", "prediction", "relation"))
+  predict <- match.arg(predict, choices = c("expectation", "link", "prediction", "response", "relation"))
   # Other names: "response", "expected", "distribution", "observations"
   if (predict == "relation") {
     message(format_message(
@@ -618,7 +636,7 @@ get_predicted.faMain <- function(x, data = NULL, ...) {
   } else if (predict == "expectation") {
     ci_type <- "confidence"
     scale <- "response"
-  } else if (predict == "prediction") {
+  } else if (predict %in% c("prediction", "response")) {
     ci_type <- "prediction"
     scale <- "response"
   }
@@ -684,6 +702,21 @@ get_predicted.faMain <- function(x, data = NULL, ...) {
 }
 
 
+# -------------------------------------------------------------------------
+.get_predict_transform_response <- function(predictions, response) {
+  predictions <- round(predictions)
+  if(is.factor(response)) {
+    predictions[predictions == 0] <- levels(response)[1]
+    predictions[predictions == 1] <- levels(response)[2]
+    predictions <- as.factor(predictions)
+    levels(predictions) <- levels(response)
+  } else {
+    predictions[predictions == 0] <- unique(response)[1]
+    predictions[predictions == 1] <- unique(response)[2]
+  }
+  predictions
+}
+
 .get_predicted_transform <- function(x,
                                      predictions,
                                      args = NULL,
@@ -712,8 +745,22 @@ get_predicted.faMain <- function(x, data = NULL, ...) {
     }
   }
 
+  # Transform to response "type"
+  if(args$predict == "response" && model_info(x)$is_binomial) {
+    response <- get_response(x)
+    ci_data[!se_col] <- as.data.frame(sapply(ci_data[!se_col], .get_predict_transform_response, response = response))
+    predictions <- .get_predict_transform_response(predictions, response = response)
+    if ("iterations" %in% names(attributes(predictions))) {
+      attr(predictions, "iterations") <- as.data.frame(sapply(attributes(predictions)$iterations, .get_predict_transform_response, response = response))
+    }
+  }
+
   list(predictions = predictions, ci_data = ci_data)
 }
+
+
+
+# -------------------------------------------------------------------------
 
 
 .get_predicted_out <- function(predictions, args = NULL, ci_data = NULL, ...) {
@@ -730,7 +777,7 @@ get_predicted.faMain <- function(x, data = NULL, ...) {
   predictions
 }
 
-# Bootstrap ---------------------------------------------------------------
+# Bootstrap ==============================================================
 
 .get_predicted_boot <- function(x,
                                 data = NULL,
@@ -771,6 +818,8 @@ get_predicted.faMain <- function(x, data = NULL, ...) {
   .get_predicted_centrality_from_draws(x, draws, ...)
 }
 
+
+# -------------------------------------------------------------------------
 
 
 .get_predicted_centrality_from_draws <- function(x,